+++ lib/pgsync/client.rb
@@ -187,6 +187,7 @@ Options:}
         o.integer "--batch-size", "batch size", default: 10000, help: false
         o.float "--sleep", "sleep", default: 0, help: false
         o.boolean "--fail-fast", "stop on the first failed table", default: false
+        o.array "--var", "pass a variable"
         o.on "-v", "--version", "print the version" do
           log PgSync::VERSION
           @exit = true