+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.5.2 (unreleased)
 
 - Added `--fail-fast` option
+- Added `--var` option
 - Improved error message
 - Fixed typo in error message