+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 
 - Added `--fail-fast` option
 - Added `--var` option
+- Automatically exclude tables when `--init` run inside Rails application
 - Improved error message
 - Fixed typo in error message
 - Deprecated `group:var` for variables
-- Automatically exclude tables when `--init` run inside Rails application
 
 ## 0.5.1 (2020-03-26)